PAAC-154: Split out model classes into separate files. Renamed test classes as appropriate.
